+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### New Features
 
+* Ability to write events to a given stream with `Eventum::Client::Events#create`. ([@enriikke][])
+
 ### Changes
 
 * Enable root `Eventum` module to access `streams` and `events` APIs. ([@enriikke][])